requirements.txt
-r tests/requirements.txt

docutils
sphinx
coverage
django-nose
django-extensions
